Update Assignment1.ipynb
I added stuff in Descriptive Statistics.
I plotted the number of reviews by ratings.
I created two new columns: one representing helpful and and the other one representing unhelpful votes that a review received.
I plotted the brands and the number of reviews that they received.
I created a new dataframe called "df_brands" where we have the unique brands, the number of reviews that they received, and decomposed this into two other columns: positive and negative feedback.
NEXT STEP: make a scatter plot with this dataframe showing the number of positive and negative feedbacks for the top 10 brands with the highest number of reviews.
DOUBT: Apple is way before other companies and I was thinking if that's because we used 'iPad' as a keyword for filtering our data.